organize files into paths
