+++ src/sprites/ParticleSprite.ts
