+++ src/sprites/RocketSprite.ts
